Connected homepage with database, and completed form functionality
