+++ templates/sidebar.html
@@ -5,18 +5,19 @@
 
 <!-- In the body of the HTML file-->
 {% block content %}
-    <div class="row">
+    <div style="margin: 0; overflow: hidden;" class="row">
 
         <!-- Static Side Bar -->
-        <nav class="nav flex-column col-sm-2 left-justified vh-100 sidebar">
+        <nav style="overflow: hidden;" class="nav flex-column col-md-2 left-justified sidebar vh-100">
             <a class="nav-link text-center sidebar-element" href="{{ url_for('routes.new_item') }}">New Item</a>
             <a class="nav-link text-center sidebar-element" href="#">Link</a>
             <a class="nav-link text-center sidebar-element" href="#">Link</a>
-            <p class="nav-link text-left sidebar-paragraph">Est ferox fluctus, cesaris.Castors sunt aonidess de azureus lanista.</p>
+            <p class="nav-link text-left sidebar-paragraph">Est ferox fluctus, cesaris.Castors sunt aonidess de azureus
+                lanista.</p>
         </nav>
 
         <!-- Dynamic Content -->
-        <div class="col-md-8">
+        <div style="overflow: hidden;" class="col-md-8">
             {% block body %}{% endblock %}
         </div>
     </div>
